Adding initial commit for libcppbor
This commit alters the binary names in the Android.bp file to avoid
conflicts with the internal copy of libcppbor in hardware/interfaces.
Once checked in, the copy will be removed and the binary names will be
reverted back to normal. This is to avoid chicken and egg build errors.

Test: tests are under tests/
Change-Id: I61355fb05d862a6809ac1e42bce714908da869b2