Don't allow implicit conversion to Array
The variadic Array constructor would implicitly make an Array object
from a single input that is convertable to an Item. This leads to
surprising behaviors wherein Arrays could be created unintentionally.

Fix this by making the variadic constructor only work for more than
one argument. Add an additional, single-item constructor that is
explicit.

Bug: 191778240
Test: cppbor_host_test_external
Change-Id: Ifadc6d3137ee285a81076a4125b6053f1d7c21c4